+++ python/python/bystro/search/index/listener.py
@@ -8,8 +8,9 @@ import os
 from ruamel.yaml import YAML
 
 from bystro.search.index.handler import go
-from bystro.search.utils.messages import IndexJobData, IndexJobResults, IndexJobCompleteMessage
-from bystro.search.utils.beanstalkd import ProgressPublisher, QueueConf, get_config_file_path, listen
+from bystro.search.utils.annotation import get_config_file_path
+from bystro.search.utils.beanstalkd import ProgressPublisher, QueueConf, listen
+from bystro.search.utils.messages import IndexJobCompleteMessage, IndexJobData, IndexJobResults
 
 TUBE = "index"
