+++ python/python/bystro/search/save/handler.py
@@ -12,15 +12,15 @@ import pathlib
 import subprocess
 import traceback
 
-from isal import igzip
 import numpy as np
-from opensearchpy import OpenSearch
 import ray
+from isal import igzip
+from opensearchpy import OpenSearch
 
-from bystro.search.utils.messages import SaveJobData
+from bystro.search.utils.annotation import AnnotationOutputs, get_delimiters
 from bystro.search.utils.beanstalkd import ProgressPublisher, get_progress_reporter
+from bystro.search.utils.messages import SaveJobData
 from bystro.search.utils.opensearch import gather_opensearch_args
-from bystro.search.utils.annotation import get_delimiters, AnnotationOutputs
 
 ray.init(ignore_reinit_error=True, address="auto")
