+++ python/python/bystro/search/utils/beanstalkd.py
@@ -5,8 +5,6 @@ import sys
 import time
 import traceback
 from collections.abc import Callable
-from glob import glob
-from os import path
 from textwrap import dedent
 from typing import Any, NamedTuple, TypeVar
 
@@ -56,20 +54,6 @@ class QueueConf(Struct):
             hosts.append(host)
             ports.append(port)
         return hosts, ports
-
-
-def get_config_file_path(config_path_base_dir: str, assembly: str, suffix: str = ".y*ml"):
-    """Get config file path"""
-    paths = glob(path.join(config_path_base_dir, assembly + suffix))
-
-    if not paths:
-        raise ValueError(f"\n\nNo config path found for the assembly {assembly}. Exiting\n\n")
-
-    if len(paths) > 1:
-        print("\n\nMore than 1 config path found, choosing first")
-
-    return paths[0]
-
 
 def default_failed_msg_fn(
     job_data: T | None, job_id: BEANSTALK_JOB_ID, err: Exception
